- Reorganización de directorios - Algunos docstrings - Nuevo nombre para clase y modulo dataframe.py
